backend/requirements.txt
Django==4.2.7
django-ninja==0.22.2
crewai==0.1.30
crewai-tools==0.0.8
openai==1.3.5
python-dotenv==1.0.0
pydantic==2.4.2

